*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@739 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/pulser.c
@@ -596,7 +596,8 @@ void p_set( long pnum, int type, Var *v )
 			vars_check( v, INT_VAR );
 			is_pulser_func( pulser_struct.set_pulse_phase_cycle,
 							"setting a pulse phase cycle" );
-			( *pulser_struct.set_pulse_phase_cycle )( pnum, v->val.lval );
+			( *pulser_struct.set_pulse_phase_cycle )( pnum, 
+													  ( int ) v->val.lval );
 			vars_pop( v );
 			break;
